requirements.txt
opencv-python>=4.13.0
torch>=2.0.0
numpy>=2.0.0
ultralytics>=8.0.0
timm>=0.9.0
pyttsx3>=2.90
Pillow>=9.0.0
git+https://github.com/openai/CLIP.git
